All of ds.h

+++ Bowtie2Java/src/com/uwb/bt2j/aligner/RandomSource.java
@@ -51,7 +51,7 @@ public class RandomSource {
 		return first | second;
 	}
 	
-	public double nextSizeT() {
+	public int nextSizeT() {
 		return nextU32();
 	}
